ui: remove npm stuff
npm was only used to install elm, but the elm package depends on an old
version of the request package with vulnerabilities. The vulnerabilites
apply to browser environments only, so they don't affect us here, but
they still make dependabot cry out every now and then.

It just occured to me that it is completely crazy to invoke the madness
that is NPM, just to download the elm compiler. So we don't do that
anymore. Just install elm as described on elm-lang.org.